Add GetChargerConnected method for org.freesmartphone.Device.PowerSupply interface
We currently don't have the possiblity to see if the charger is still connected to the
device when the battery is fully charged. In this case the power supply status is FULL and
we can't see if the charger is there. For this case the new method GetChargerConnected
gives the user the possiblity to see if the charger is connected to the device.